+++ packages/core-js-sdk/src/sdk/types.ts
@@ -50,6 +50,9 @@ export type UserResponse = User & {
   roleNames?: string[];
   userTenants?: UserTenant[];
   createTime: number;
+  totp: boolean;
+  saml: boolean;
+  oauth?: Record<string, boolean>;
 };
 
 /** A tenant association mapping  */
@@ -220,6 +223,7 @@ export type Options = {
   deviceInfo?: DeviceInfo;
   lastAuth?: LastAuth;
   redirectAuth?: RedirectAuth;
+  oidcIdpStateId?: string;
 };
 
 export type ResponseData = Record<string, any>;